Added some comments, and merged some contributed comments from Matthias Berth.
I prefer to lazy load my accessors rather than using initialize.  Doing so requires that in general you never access fields directly, however, for collection, I want to return a copy of the collection so that a client can't modify it and change an objects state indirectly.  

comments
	^ (comments
		ifNil: [comments := OrderedCollection new]) copy

This pattern has the concequence that in my add and remove accessors for that item, I need to access the inst var directly, which requires a hit to the accessor first to ensure any lazy initialization has taken place.  

addComment: aComment 
	self comments.
	^ comments add: aComment

While it may seem like a bit of extra work, it's not since I have my IDE generate those accessors for me.